==> script/Analysis 2021/7-GLMM.Rmd <==

```{r include=FALSE}
library(tidyverse)
library(lme4)
library(car)
library(multcomp)
library(readxl)
library(MuMIn)
library(here)
library(flextable)
```


```{r include=FALSE}

M.data <- read_excel(here("data/clean/for analysis_1521.xlsx"),
                     sheet=1) %>% 
  filter(analysis %in% "Y") %>% 
  mutate(TypeName.1 = case_when(
    str_detect(TypeName ,"混") ~ "mixed",
    str_detect(TypeName ,"竹林") ~  "Bamboo",
    str_detect(TypeName ,"闊葉樹林型") ~  "broad-leaved",
    str_detect(TypeName ,"針葉樹林型") ~  "coniferous",
         TRUE ~ "Not forest"
  )) %>% 
  mutate_at(c("Year", "Survey", "Point", "Macaca_sur",
              "Month", "Day", "Distance", "julian.D"),as.numeric) %>% 
  mutate_at(c("TypeName.1", "Site_N", "Region2"),as.factor)   


```

#---------------------------------------------------------------------


```{r include=FALSE}
M.data <- M.data %>% 
  mutate(Macaca_sur = ifelse(is.na(Macaca_sur), 0, Macaca_sur)) %>% 
  mutate(Year.re = Year - min(Year) + 1)

```


#==============================================

```{r include=FALSE}
df <- 
  M.data 
```

#-------------------------------------------
```{r eval=FALSE, include=FALSE}
allFit(glmer(Macaca_sur ~ TypeName.1  + Year.re + Altitude + julian.D +  Region2 + (1|Site_N), 
             family = binomial, data = df))   #嘗試使用一系列優化程序重新擬合glmer模型


```

```{r include=FALSE}
df$Altitude.1 <-  scale(df$Altitude,scale =T)
df$julian.D.1 <-  scale(df$julian.D,scale =T)
```


```{r}
m1 <- glmer(Macaca_sur ~  TypeName.1 + Year.re + Altitude.1 + julian.D.1 +  Region2 + (1|Site_N), 
            family = binomial, data = df,
            control = glmerControl(optimizer = "bobyqa"))
```


```{r}
Anova(m1)

```

```{r}
df %>% 
  filter(Macaca_sur ==1) %>% 
  dim
```


```{r include=FALSE}
summary(m1)


```

```{r include=FALSE}
#AICc==============================================
options(na.action = "na.fail")
d1<- dredge(m1, trace = F)  
#trace = T的話，是把過程都show出來，看有幾種組合。
```


```{r include=FALSE}
d1[d1$delta<2,] 

```


```{r include=FALSE}
summary(model.avg(d1))
summary(model.avg(d1, subset = delta < 2))
```



```{r}
sw(d1)
```


```{r include=FALSE}
sw(model.avg(d1, subset = delta < 2))

```

```{r include=FALSE}

m2 <- glmer(Macaca_sur ~ Altitude.1 + Year.re  +  julian.D.1 +  Region2 + (1|Site_N), 
            family = binomial, data = df,
            control = glmerControl(optimizer = "bobyqa"))
```

#anova table

```{r}
Anova(m2)
```


```{r include=FALSE}
summary(m2)
```


```{r}
summary(glht(m2, linfct = mcp(Region2 = "Tukey")))
```


```{r}
summary(glht(m2, linfct = c("Year.re = 0",
                            "Altitude.1 = 0",
                            "julian.D.1 = 0")))
```


```{r eval=FALSE, include=FALSE}
par(mai=c(1,1.5,1,1))
glht(m2, linfct = mcp(Region2 = "Tukey")) %>% plot


par(mai=c(1,1.5,1,1))
glht(m2, linfct = mcp(Region2 = "Tukey")) %>% 
  cld(.,level=0.05) %>% 
plot(.,col="lightgrey")
```


```{r eval=FALSE, include=FALSE}
#-----
ND <- 
df %>% 
  dplyr::select(Year.re, Region2, julian.D.1, Year, julian.D, Site_N, Altitude.1, Altitude) 

ND$pred= predict(m2, ND,re.form=NA,
               type="response")

ND %>% 
  dplyr::group_by(Year) %>% 
  dplyr::summarise(Mean = mean(pred), 
                   DL = quantile(pred,0.025),
                   UL = quantile(pred,0.975)) %>% 
  ggplot(., aes(x= Year, y = Mean))+
  geom_ribbon(aes(ymin = DL, ymax = UL), fill = "grey80")+
  geom_point()+
  geom_line(lty = 1)+
  theme_bw()
  






```
